+++ src/server/services/govee/own-govee-device.ts
@@ -179,8 +179,7 @@ export class OwnGoveeDevice implements iLedRgbCct, iTemporaryDisableAutomatic {
   }
 
   public restoreTargetAutomaticValue(c: RestoreTargetAutomaticValueCommand): void {
-    this.log(LogLevel.Debug, `Restore Target Automatic value`);
-    this.setActuator(new ActuatorSetStateCommand(c, this.targetAutomaticState, 'Restore Target Automatic value'));
+    this.setActuator(new ActuatorSetStateCommand(c, this.targetAutomaticState));
   }
 
   public persist(): void {